[Agent] extract memory storage provider util
